functional_tests/requirements.txt
annotated-types==0.7.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
flake8==7.3.0
markupsafe==3.0.2
packaging==25.0
pdoc3==0.11.6
pip==25.1.1
playwright==1.53.0
pytest-playwright==0.7.0
install_playwright==0.1.0
pluggy==1.6.0
pydantic-core==2.27.2
pydantic-settings==2.10.1
python-dotenv==1.1.0
python-hpilo==4.4.3
pytest==8.4.0
pytest-check>=2.2.0
pytest-cov>=4.0.0
pytest-dependency>=0.5.1
pytest-order>=1.0.0
pytest-md-report>=0.6.2
pytest-reportportal==5.5.1
pytest-xdist>=3.3.1
requests==2.32.4
reportportal-client==5.6.5
setuptools==80.9.0
typing-extensions==4.12.2
uv==0.4.13
werkzeug==3.1.3
waiting==1.5.0
hpe-pc-automation-lib>=2.202.0
hpe-glcp-automation-lib>=2.1849.0
#hpe-morpheus-automation-libs==8.0.7.dev29+custom
hpe-morpheus-automation-libs>=8.0.9.dev109